Routing and re structurize component
